General sanity and logging updates
- Clarify a few comments better
- Change how RedirectUser gets loaded
- Added logging to allow users to follow execution
- Made the bottle sample a bit more explicit in terms of giving back a
  result
