+++ samples/bottle-sample.py
@@ -1,4 +1,6 @@
-from bottle import run, request, route
+from bottle import run, request, route, sys, os
+# Not ideal, but for the sake of an example
+sys.path.append(os.path.join(os.path.dirname(os.path.dirname(__file__)), "src"))
 from steamsignin import SteamSignIn
 
 @route('/')
@@ -21,7 +23,15 @@ def process():
 	steamLogin = SteamSignIn()
 	steamID = steamLogin.ValidateResults(returnData)
 
-	print(steamID)
+	print("SteamID returned was: ", steamID)
 
+	if steamID is not False:
+		return "We logged in successfully!<br />SteamID: {0}".format(steamID)
+	else:
+		return "Failed to log in, bad details?"
+
+	# At this point, redirect the user to a friendly URL
+
+	
 if __name__ == '__main__':
     run(host='localhost', port=8080, debug = True, reloader = False)
